🔧 Settings : getQueryClient와 Query Provider 파일 분할
